+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -11,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Key-Value store for instance settings
 - Allow for Re-extractions of projects to a database, see [ExecuteFullExtractionToDatabaseMSSql](Documentation\DataExtractions\ExecuteFullExtractionToDatabaseMSSql.md)
+- When cloning an ExtractionConfiguration with a depricated catalogue, the GUI will ask if you want to replace the depricated catalogue with the known replacement
 
 ## [8.1.6] - 2024-05-27